refactor:  Refactor SpotMap component to separate concerns into individual components and hooks
