+++ apps/web/src/features/spots/config/index.ts
@@ -1,0 +1,1 @@
+export { mapOptions } from './mapOptions';
