+++ apps/web/src/features/spots/model/index.ts
@@ -1,0 +1,2 @@
+export { default as useMarkerVisibility } from './useMarkerVisibility';
+export { default as useMapState } from './useMapState';
